update queue module
